refactor(API): loosen restrictions on `delegate`/`set_delegate` API
ContractInstance -> anything that can be converted to AddressType
