+++ src/ape/api/accounts.py
@@ -478,7 +478,7 @@ class AccountAPI(BaseInterfaceModel, BaseAddress):
         nonce = self.nonce if nonce is None else nonce
         return ecosystem.get_deployment_address(self.address, nonce)
 
-    def set_delegate(self, contract: "ContractInstance", **txn_kwargs):
+    def set_delegate(self, contract: Union[BaseAddress, AddressType, str], **txn_kwargs):
         """
         Have the account class override the value of its ``delegate``. For plugins that support
         this feature, the way they choose to handle it can vary. For example, it could be a call to
@@ -523,11 +523,11 @@ class AccountAPI(BaseInterfaceModel, BaseAddress):
     @contextmanager
     def delegate_to(
         self,
-        new_delegate: "ContractInstance",
+        new_delegate: Union[BaseAddress, AddressType, str],
         set_txn_kwargs: Optional[dict] = None,
         reset_txn_kwargs: Optional[dict] = None,
         **txn_kwargs,
-    ) -> Iterator["ContractInstance"]:
+    ) -> Iterator[BaseAddress]:
         """
         Temporarily override the value of ``delegate`` for the account inside of a context manager,
         and yields a contract instance object whose interface matches that of ``new_delegate``.
@@ -566,7 +566,12 @@ class AccountAPI(BaseInterfaceModel, BaseAddress):
 
         # This is helpful for using it immediately to send things as self
         with self.account_manager.use_sender(self):
-            yield ContractInstance(self.address, contract_type=new_delegate.contract_type)
+            if isinstance(new_delegate, ContractInstance):
+                # NOTE: Do not cache this
+                yield ContractInstance(self.address, contract_type=new_delegate.contract_type)
+
+            else:
+                yield self
 
         reset_txn_kwargs = {**txn_kwargs, **(reset_txn_kwargs or {})}
         if existing_delegate:
