+++ src/ape/managers/chain.py
@@ -9,7 +9,7 @@ import pandas as pd
 from rich.box import SIMPLE
 from rich.table import Table
 
-from ape.api.address import BaseAddress
+from ape.api.address import Address, BaseAddress
 from ape.api.providers import BlockAPI
 from ape.api.query import (
     AccountTransactionQuery,
@@ -22,6 +22,7 @@ from ape.exceptions import (
     APINotImplementedError,
     BlockNotFoundError,
     ChainError,
+    ContractNotFoundError,
     ProviderNotConnectedError,
     QueryEngineError,
     TransactionNotFoundError,
@@ -37,7 +38,6 @@ from ape.utils.misc import ZERO_ADDRESS, is_evm_precompile, is_zero_hex, log_ins
 if TYPE_CHECKING:
     from rich.console import Console as RichConsole
 
-    from ape.contracts import ContractInstance
     from ape.types import BlockID, ContractCode, GasReport, SnapshotID, SourceTraceback
 
 
@@ -1018,7 +1018,7 @@ class ChainManager(BaseManager):
         self._code[network.ecosystem.name][network.name][address] = code
         return code
 
-    def get_delegate(self, address: AddressType) -> Optional["ContractInstance"]:
+    def get_delegate(self, address: AddressType) -> Optional[BaseAddress]:
         ecosystem = self.provider.network.ecosystem
 
         if not (proxy_info := ecosystem.get_proxy_info(address)):
@@ -1026,4 +1026,9 @@ class ChainManager(BaseManager):
 
         # NOTE: Do this every time because it can change?
         self.contracts.cache_proxy_info(address, proxy_info)
-        return self.contracts.instance_at(proxy_info.target)
+
+        try:
+            return self.contracts.instance_at(proxy_info.target)
+
+        except ContractNotFoundError:
+            return Address(proxy_info.target)
